Move share_url validations into AR concern and move their tests to RSpec shared_examples
